MCP/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-mcp-server"
version = "0.1.0"
description = "MCP server that provides stock momentum analysis and AI recommendations"
readme = "README.md"
authors = [{name = "Your Name"}]
requires-python = ">=3.9"
dependencies = [
  "mcp[cli]>=1.1.0",
  "pandas>=2.0.0",
  "numpy>=1.24.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "google-generativeai>=0.8.0",
  "openai>=1.37.0"
]

[project.scripts]
stock-mcp = "stock_mcp_server.server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["stock_mcp_server*"]
